requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
crypto==1.4.1
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
Naked==0.1.31
psycopg2==2.8.5
pycryptodome==3.9.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
sqlparse==0.3.1
urllib3==1.25.9
